Refactor: engine関連のstoreを分離
