+++ src/store/index.ts
@@ -25,6 +25,7 @@ import { dictionaryStoreState, dictionaryStore } from "./dictionary";
 import { proxyStore, proxyStoreState } from "./proxy";
 import { createPartialStore } from "./vuex";
 import { DefaultStyleId } from "@/type/preload";
+import { engineStoreState, engineStore } from "./engine";
 
 export const storeKey: InjectionKey<
   Store<State, AllGetters, AllActions, AllMutations>
@@ -292,6 +293,7 @@ export const store = createStore<State, AllGetters, AllActions, AllMutations>({
     ...uiStoreState,
     ...audioStoreState,
     ...commandStoreState,
+    ...engineStoreState,
     ...projectStoreState,
     ...settingStoreState,
     ...audioCommandStoreState,
@@ -305,6 +307,7 @@ export const store = createStore<State, AllGetters, AllActions, AllMutations>({
     ...uiStore.getters,
     ...audioStore.getters,
     ...commandStore.getters,
+    ...engineStore.getters,
     ...projectStore.getters,
     ...settingStore.getters,
     ...presetStore.getters,
@@ -318,6 +321,7 @@ export const store = createStore<State, AllGetters, AllActions, AllMutations>({
     ...uiStore.mutations,
     ...audioStore.mutations,
     ...commandStore.mutations,
+    ...engineStore.mutations,
     ...projectStore.mutations,
     ...settingStore.mutations,
     ...audioCommandStore.mutations,
@@ -330,6 +334,7 @@ export const store = createStore<State, AllGetters, AllActions, AllMutations>({
   actions: {
     ...uiStore.actions,
     ...audioStore.actions,
+    ...engineStore.actions,
     ...commandStore.actions,
     ...projectStore.actions,
     ...settingStore.actions,
